+++ plugins/module_utils/acme/__init__.py
@@ -27,8 +27,6 @@ from ansible.module_utils.urls import fetch_url
 from ansible.module_utils.six.moves.urllib.parse import unquote
 from ansible.module_utils._text import to_native, to_text, to_bytes
 
-from ansible_collections.community.crypto.plugins.module_utils.compat import ipaddress as compat_ipaddress
-
 from ansible_collections.community.crypto.plugins.module_utils.acme.errors import ModuleFailException
 
 from ansible_collections.community.crypto.plugins.module_utils.acme.io import (
@@ -42,358 +40,18 @@ from ansible_collections.community.crypto.plugins.module_utils.acme.utils import
     process_links,
 )
 
-try:
-    import cryptography
-    import cryptography.hazmat.backends
-    import cryptography.hazmat.primitives.hashes
-    import cryptography.hazmat.primitives.hmac
-    import cryptography.hazmat.primitives.asymmetric.ec
-    import cryptography.hazmat.primitives.asymmetric.padding
-    import cryptography.hazmat.primitives.asymmetric.rsa
-    import cryptography.hazmat.primitives.asymmetric.utils
-    import cryptography.hazmat.primitives.serialization
-    import cryptography.x509
-    import cryptography.x509.oid
-    from distutils.version import LooseVersion
-    CRYPTOGRAPHY_VERSION = cryptography.__version__
-    HAS_CURRENT_CRYPTOGRAPHY = (LooseVersion(CRYPTOGRAPHY_VERSION) >= LooseVersion('1.5'))
-    if HAS_CURRENT_CRYPTOGRAPHY:
-        _cryptography_backend = cryptography.hazmat.backends.default_backend()
-except Exception as dummy:
-    HAS_CURRENT_CRYPTOGRAPHY = False
+from ansible_collections.community.crypto.plugins.module_utils.acme.backend_openssl_cli import (
+    OpenSSLCLIBackend,
+)
 
+from ansible_collections.community.crypto.plugins.module_utils.acme.backend_cryptography import (
+    CryptographyBackend,
+    CRYPTOGRAPHY_VERSION,
+)
 
-def _parse_key_openssl(openssl_binary, module, key_file=None, key_content=None):
-    '''
-    Parses an RSA or Elliptic Curve key file in PEM format and returns a pair
-    (error, key_data).
-    '''
-    # If key_file isn't given, but key_content, write that to a temporary file
-    if key_file is None:
-        fd, tmpsrc = tempfile.mkstemp()
-        module.add_cleanup_file(tmpsrc)  # Ansible will delete the file on exit
-        f = os.fdopen(fd, 'wb')
-        try:
-            f.write(key_content.encode('utf-8'))
-            key_file = tmpsrc
-        except Exception as err:
-            try:
-                f.close()
-            except Exception as dummy:
-                pass
-            raise ModuleFailException("failed to create temporary content file: %s" % to_native(err), exception=traceback.format_exc())
-        f.close()
-    # Parse key
-    account_key_type = None
-    with open(key_file, "rt") as f:
-        for line in f:
-            m = re.match(r"^\s*-{5,}BEGIN\s+(EC|RSA)\s+PRIVATE\s+KEY-{5,}\s*$", line)
-            if m is not None:
-                account_key_type = m.group(1).lower()
-                break
-    if account_key_type is None:
-        # This happens for example if openssl_privatekey created this key
-        # (as opposed to the OpenSSL binary). For now, we assume this is
-        # an RSA key.
-        # FIXME: add some kind of auto-detection
-        account_key_type = "rsa"
-    if account_key_type not in ("rsa", "ec"):
-        return 'unknown key type "%s"' % account_key_type, {}
+from ansible_collections.community.crypto.plugins.module_utils.acme.backend_cryptography import HAS_CURRENT_CRYPTOGRAPHY as _ORIGINAL_HAS_CURRENT_CRYPTOGRAPHY
 
-    openssl_keydump_cmd = [openssl_binary, account_key_type, "-in", key_file, "-noout", "-text"]
-    dummy, out, dummy = module.run_command(openssl_keydump_cmd, check_rc=True)
-
-    if account_key_type == 'rsa':
-        pub_hex, pub_exp = re.search(
-            r"modulus:\n\s+00:([a-f0-9\:\s]+?)\npublicExponent: ([0-9]+)",
-            to_text(out, errors='surrogate_or_strict'), re.MULTILINE | re.DOTALL).groups()
-        pub_exp = "{0:x}".format(int(pub_exp))
-        if len(pub_exp) % 2:
-            pub_exp = "0{0}".format(pub_exp)
-
-        return None, {
-            'key_file': key_file,
-            'type': 'rsa',
-            'alg': 'RS256',
-            'jwk': {
-                "kty": "RSA",
-                "e": nopad_b64(binascii.unhexlify(pub_exp.encode("utf-8"))),
-                "n": nopad_b64(binascii.unhexlify(re.sub(r"(\s|:)", "", pub_hex).encode("utf-8"))),
-            },
-            'hash': 'sha256',
-        }
-    elif account_key_type == 'ec':
-        pub_data = re.search(
-            r"pub:\s*\n\s+04:([a-f0-9\:\s]+?)\nASN1 OID: (\S+)(?:\nNIST CURVE: (\S+))?",
-            to_text(out, errors='surrogate_or_strict'), re.MULTILINE | re.DOTALL)
-        if pub_data is None:
-            return 'cannot parse elliptic curve key', {}
-        pub_hex = binascii.unhexlify(re.sub(r"(\s|:)", "", pub_data.group(1)).encode("utf-8"))
-        asn1_oid_curve = pub_data.group(2).lower()
-        nist_curve = pub_data.group(3).lower() if pub_data.group(3) else None
-        if asn1_oid_curve == 'prime256v1' or nist_curve == 'p-256':
-            bits = 256
-            alg = 'ES256'
-            hashalg = 'sha256'
-            point_size = 32
-            curve = 'P-256'
-        elif asn1_oid_curve == 'secp384r1' or nist_curve == 'p-384':
-            bits = 384
-            alg = 'ES384'
-            hashalg = 'sha384'
-            point_size = 48
-            curve = 'P-384'
-        elif asn1_oid_curve == 'secp521r1' or nist_curve == 'p-521':
-            # Not yet supported on Let's Encrypt side, see
-            # https://github.com/letsencrypt/boulder/issues/2217
-            bits = 521
-            alg = 'ES512'
-            hashalg = 'sha512'
-            point_size = 66
-            curve = 'P-521'
-        else:
-            return 'unknown elliptic curve: %s / %s' % (asn1_oid_curve, nist_curve), {}
-        num_bytes = (bits + 7) // 8
-        if len(pub_hex) != 2 * num_bytes:
-            return 'bad elliptic curve point (%s / %s)' % (asn1_oid_curve, nist_curve), {}
-        return None, {
-            'key_file': key_file,
-            'type': 'ec',
-            'alg': alg,
-            'jwk': {
-                "kty": "EC",
-                "crv": curve,
-                "x": nopad_b64(pub_hex[:num_bytes]),
-                "y": nopad_b64(pub_hex[num_bytes:]),
-            },
-            'hash': hashalg,
-            'point_size': point_size,
-        }
-
-
-def _create_mac_key_openssl(openssl_bin, module, alg, key):
-    if alg == 'HS256':
-        hashalg = 'sha256'
-        hashbytes = 32
-    elif alg == 'HS384':
-        hashalg = 'sha384'
-        hashbytes = 48
-    elif alg == 'HS512':
-        hashalg = 'sha512'
-        hashbytes = 64
-    else:
-        raise ModuleFailException('Unsupported MAC key algorithm for OpenSSL backend: {0}'.format(alg))
-    key_bytes = base64.urlsafe_b64decode(key)
-    if len(key_bytes) < hashbytes:
-        raise ModuleFailException(
-            '{0} key must be at least {1} bytes long (after Base64 decoding)'.format(alg, hashbytes))
-    return {
-        'type': 'hmac',
-        'alg': alg,
-        'jwk': {
-            'kty': 'oct',
-            'k': key,
-        },
-        'hash': hashalg,
-    }
-
-
-def _sign_request_openssl(openssl_binary, module, payload64, protected64, key_data):
-    sign_payload = "{0}.{1}".format(protected64, payload64).encode('utf8')
-    if key_data['type'] == 'hmac':
-        hex_key = to_native(binascii.hexlify(base64.urlsafe_b64decode(key_data['jwk']['k'])))
-        cmd_postfix = ["-mac", "hmac", "-macopt", "hexkey:{0}".format(hex_key), "-binary"]
-    else:
-        cmd_postfix = ["-sign", key_data['key_file']]
-    openssl_sign_cmd = [openssl_binary, "dgst", "-{0}".format(key_data['hash'])] + cmd_postfix
-
-    dummy, out, dummy = module.run_command(openssl_sign_cmd, data=sign_payload, check_rc=True, binary_data=True)
-
-    if key_data['type'] == 'ec':
-        dummy, der_out, dummy = module.run_command(
-            [openssl_binary, "asn1parse", "-inform", "DER"],
-            data=out, binary_data=True)
-        expected_len = 2 * key_data['point_size']
-        sig = re.findall(
-            r"prim:\s+INTEGER\s+:([0-9A-F]{1,%s})\n" % expected_len,
-            to_text(der_out, errors='surrogate_or_strict'))
-        if len(sig) != 2:
-            raise ModuleFailException(
-                "failed to generate Elliptic Curve signature; cannot parse DER output: {0}".format(
-                    to_text(der_out, errors='surrogate_or_strict')))
-        sig[0] = (expected_len - len(sig[0])) * '0' + sig[0]
-        sig[1] = (expected_len - len(sig[1])) * '0' + sig[1]
-        out = binascii.unhexlify(sig[0]) + binascii.unhexlify(sig[1])
-
-    return {
-        "protected": protected64,
-        "payload": payload64,
-        "signature": nopad_b64(to_bytes(out)),
-    }
-
-
-if sys.version_info[0] >= 3:
-    # Python 3 (and newer)
-    def _count_bytes(n):
-        return (n.bit_length() + 7) // 8 if n > 0 else 0
-
-    def _convert_int_to_bytes(count, no):
-        return no.to_bytes(count, byteorder='big')
-
-    def _pad_hex(n, digits):
-        res = hex(n)[2:]
-        if len(res) < digits:
-            res = '0' * (digits - len(res)) + res
-        return res
-else:
-    # Python 2
-    def _count_bytes(n):
-        if n <= 0:
-            return 0
-        h = '%x' % n
-        return (len(h) + 1) // 2
-
-    def _convert_int_to_bytes(count, n):
-        h = '%x' % n
-        if len(h) > 2 * count:
-            raise Exception('Number {1} needs more than {0} bytes!'.format(count, n))
-        return ('0' * (2 * count - len(h)) + h).decode('hex')
-
-    def _pad_hex(n, digits):
-        h = '%x' % n
-        if len(h) < digits:
-            h = '0' * (digits - len(h)) + h
-        return h
-
-
-def _parse_key_cryptography(module, key_file=None, key_content=None):
-    '''
-    Parses an RSA or Elliptic Curve key file in PEM format and returns a pair
-    (error, key_data).
-    '''
-    # If key_content isn't given, read key_file
-    if key_content is None:
-        key_content = read_file(key_file)
-    else:
-        key_content = to_bytes(key_content)
-    # Parse key
-    try:
-        key = cryptography.hazmat.primitives.serialization.load_pem_private_key(key_content, password=None, backend=_cryptography_backend)
-    except Exception as e:
-        return 'error while loading key: {0}'.format(e), None
-    if isinstance(key, cryptography.hazmat.primitives.asymmetric.rsa.RSAPrivateKey):
-        pk = key.public_key().public_numbers()
-        return None, {
-            'key_obj': key,
-            'type': 'rsa',
-            'alg': 'RS256',
-            'jwk': {
-                "kty": "RSA",
-                "e": nopad_b64(_convert_int_to_bytes(_count_bytes(pk.e), pk.e)),
-                "n": nopad_b64(_convert_int_to_bytes(_count_bytes(pk.n), pk.n)),
-            },
-            'hash': 'sha256',
-        }
-    elif isinstance(key, cryptography.hazmat.primitives.asymmetric.ec.EllipticCurvePrivateKey):
-        pk = key.public_key().public_numbers()
-        if pk.curve.name == 'secp256r1':
-            bits = 256
-            alg = 'ES256'
-            hashalg = 'sha256'
-            point_size = 32
-            curve = 'P-256'
-        elif pk.curve.name == 'secp384r1':
-            bits = 384
-            alg = 'ES384'
-            hashalg = 'sha384'
-            point_size = 48
-            curve = 'P-384'
-        elif pk.curve.name == 'secp521r1':
-            # Not yet supported on Let's Encrypt side, see
-            # https://github.com/letsencrypt/boulder/issues/2217
-            bits = 521
-            alg = 'ES512'
-            hashalg = 'sha512'
-            point_size = 66
-            curve = 'P-521'
-        else:
-            return 'unknown elliptic curve: {0}'.format(pk.curve.name), {}
-        num_bytes = (bits + 7) // 8
-        return None, {
-            'key_obj': key,
-            'type': 'ec',
-            'alg': alg,
-            'jwk': {
-                "kty": "EC",
-                "crv": curve,
-                "x": nopad_b64(_convert_int_to_bytes(num_bytes, pk.x)),
-                "y": nopad_b64(_convert_int_to_bytes(num_bytes, pk.y)),
-            },
-            'hash': hashalg,
-            'point_size': point_size,
-        }
-    else:
-        return 'unknown key type "{0}"'.format(type(key)), {}
-
-
-def _create_mac_key_cryptography(module, alg, key):
-    if alg == 'HS256':
-        hashalg = cryptography.hazmat.primitives.hashes.SHA256
-        hashbytes = 32
-    elif alg == 'HS384':
-        hashalg = cryptography.hazmat.primitives.hashes.SHA384
-        hashbytes = 48
-    elif alg == 'HS512':
-        hashalg = cryptography.hazmat.primitives.hashes.SHA512
-        hashbytes = 64
-    else:
-        raise ModuleFailException('Unsupported MAC key algorithm for cryptography backend: {0}'.format(alg))
-    key_bytes = base64.urlsafe_b64decode(key)
-    if len(key_bytes) < hashbytes:
-        raise ModuleFailException(
-            '{0} key must be at least {1} bytes long (after Base64 decoding)'.format(alg, hashbytes))
-    return {
-        'mac_obj': lambda: cryptography.hazmat.primitives.hmac.HMAC(
-            key_bytes,
-            hashalg(),
-            _cryptography_backend),
-        'type': 'hmac',
-        'alg': alg,
-        'jwk': {
-            'kty': 'oct',
-            'k': key,
-        },
-    }
-
-
-def _sign_request_cryptography(module, payload64, protected64, key_data):
-    sign_payload = "{0}.{1}".format(protected64, payload64).encode('utf8')
-    if 'mac_obj' in key_data:
-        mac = key_data['mac_obj']()
-        mac.update(sign_payload)
-        signature = mac.finalize()
-    elif isinstance(key_data['key_obj'], cryptography.hazmat.primitives.asymmetric.rsa.RSAPrivateKey):
-        padding = cryptography.hazmat.primitives.asymmetric.padding.PKCS1v15()
-        hashalg = cryptography.hazmat.primitives.hashes.SHA256
-        signature = key_data['key_obj'].sign(sign_payload, padding, hashalg())
-    elif isinstance(key_data['key_obj'], cryptography.hazmat.primitives.asymmetric.ec.EllipticCurvePrivateKey):
-        if key_data['hash'] == 'sha256':
-            hashalg = cryptography.hazmat.primitives.hashes.SHA256
-        elif key_data['hash'] == 'sha384':
-            hashalg = cryptography.hazmat.primitives.hashes.SHA384
-        elif key_data['hash'] == 'sha512':
-            hashalg = cryptography.hazmat.primitives.hashes.SHA512
-        ecdsa = cryptography.hazmat.primitives.asymmetric.ec.ECDSA(hashalg())
-        r, s = cryptography.hazmat.primitives.asymmetric.utils.decode_dss_signature(key_data['key_obj'].sign(sign_payload, ecdsa))
-        rr = _pad_hex(r, 2 * key_data['point_size'])
-        ss = _pad_hex(s, 2 * key_data['point_size'])
-        signature = binascii.unhexlify(rr) + binascii.unhexlify(ss)
-
-    return {
-        "protected": protected64,
-        "payload": payload64,
-        "signature": nopad_b64(signature),
-    }
+HAS_CURRENT_CRYPTOGRAPHY = _ORIGINAL_HAS_CURRENT_CRYPTOGRAPHY
 
 
 def _assert_fetch_url_success(response, info, allow_redirect=False, allow_client_error=True, allow_server_error=True):
@@ -457,6 +115,7 @@ class ACMEAccount(object):
         self._debug = False
 
         self.module = module
+        self.backend = get_backend(module)
         self.version = module.params['acme_version']
         # account_key path and content are mutually exclusive
         self.key = module.params['account_key_src']
@@ -499,10 +158,7 @@ class ACMEAccount(object):
         '''
         if key_file is None and key_content is None:
             raise AssertionError('One of key_file and key_content must be specified!')
-        if HAS_CURRENT_CRYPTOGRAPHY:
-            return _parse_key_cryptography(self.module, key_file, key_content)
-        else:
-            return _parse_key_openssl(self._openssl_bin, self.module, key_file, key_content)
+        return self.backend.parse_key(key_file, key_content)
 
     def sign_request(self, protected, payload, key_data, encode_payload=True):
         try:
@@ -518,17 +174,7 @@ class ACMEAccount(object):
         except Exception as e:
             raise ModuleFailException("Failed to encode payload / headers as JSON: {0}".format(e))
 
-        if HAS_CURRENT_CRYPTOGRAPHY:
-            return _sign_request_cryptography(self.module, payload64, protected64, key_data)
-        else:
-            return _sign_request_openssl(self._openssl_bin, self.module, payload64, protected64, key_data)
-
-    def _create_mac_key(self, alg, key):
-        '''Create a MAC key.'''
-        if HAS_CURRENT_CRYPTOGRAPHY:
-            return _create_mac_key_cryptography(self.module, alg, key)
-        else:
-            return _create_mac_key_openssl(self._openssl_bin, self.module, alg, key)
+        return self.backend.sign(payload64, protected64, key_data)
 
     def _log(self, msg, data=None):
         '''
@@ -717,7 +363,7 @@ class ACMEAccount(object):
                         'url': url,
                     },
                     self.jwk,
-                    self._create_mac_key(external_account_binding['alg'], external_account_binding['key'])
+                    self.backend.create_mac_key(external_account_binding['alg'], external_account_binding['key'])
                 )
             elif self.directory['meta'].get('externalAccountRequired') and allow_creation:
                 raise ModuleFailException(
@@ -880,94 +526,6 @@ class ACMEAccount(object):
         return True, account_data
 
 
-def _normalize_ip(ip):
-    try:
-        return to_native(compat_ipaddress.ip_address(to_text(ip)).compressed)
-    except ValueError:
-        # We don't want to error out on something IPAddress() can't parse
-        return ip
-
-
-def openssl_get_csr_identifiers(openssl_binary, module, csr_filename, csr_content=None):
-    '''
-    Return a set of requested identifiers (CN and SANs) for the CSR.
-    Each identifier is a pair (type, identifier), where type is either
-    'dns' or 'ip'.
-    '''
-    filename = csr_filename
-    data = None
-    if csr_content is not None:
-        filename = '-'
-        data = csr_content.encode('utf-8')
-
-    openssl_csr_cmd = [openssl_binary, "req", "-in", filename, "-noout", "-text"]
-    dummy, out, dummy = module.run_command(openssl_csr_cmd, data=data, check_rc=True)
-
-    identifiers = set([])
-    common_name = re.search(r"Subject:.* CN\s?=\s?([^\s,;/]+)", to_text(out, errors='surrogate_or_strict'))
-    if common_name is not None:
-        identifiers.add(('dns', common_name.group(1)))
-    subject_alt_names = re.search(
-        r"X509v3 Subject Alternative Name: (?:critical)?\n +([^\n]+)\n",
-        to_text(out, errors='surrogate_or_strict'), re.MULTILINE | re.DOTALL)
-    if subject_alt_names is not None:
-        for san in subject_alt_names.group(1).split(", "):
-            if san.lower().startswith("dns:"):
-                identifiers.add(('dns', san[4:]))
-            elif san.lower().startswith("ip:"):
-                identifiers.add(('ip', _normalize_ip(san[3:])))
-            elif san.lower().startswith("ip address:"):
-                identifiers.add(('ip', _normalize_ip(san[11:])))
-            else:
-                raise ModuleFailException('Found unsupported SAN identifier "{0}"'.format(san))
-    return identifiers
-
-
-def cryptography_get_csr_identifiers(module, csr_filename, csr_content=None):
-    '''
-    Return a set of requested identifiers (CN and SANs) for the CSR.
-    Each identifier is a pair (type, identifier), where type is either
-    'dns' or 'ip'.
-    '''
-    identifiers = set([])
-    if csr_content is None:
-        csr_content = read_file(csr_filename)
-    else:
-        csr_content = to_bytes(csr_content)
-    csr = cryptography.x509.load_pem_x509_csr(csr_content, _cryptography_backend)
-    for sub in csr.subject:
-        if sub.oid == cryptography.x509.oid.NameOID.COMMON_NAME:
-            identifiers.add(('dns', sub.value))
-    for extension in csr.extensions:
-        if extension.oid == cryptography.x509.oid.ExtensionOID.SUBJECT_ALTERNATIVE_NAME:
-            for name in extension.value:
-                if isinstance(name, cryptography.x509.DNSName):
-                    identifiers.add(('dns', name.value))
-                elif isinstance(name, cryptography.x509.IPAddress):
-                    identifiers.add(('ip', name.value.compressed))
-                else:
-                    raise ModuleFailException('Found unsupported SAN identifier {0}'.format(name))
-    return identifiers
-
-
-def cryptography_get_cert_days(module, cert_file, now=None):
-    '''
-    Return the days the certificate in cert_file remains valid and -1
-    if the file was not found. If cert_file contains more than one
-    certificate, only the first one will be considered.
-    '''
-    if not os.path.exists(cert_file):
-        return -1
-
-    try:
-        cert = cryptography.x509.load_pem_x509_certificate(read_file(cert_file), _cryptography_backend)
-    except Exception as e:
-        raise ModuleFailException('Cannot parse certificate {0}: {1}'.format(cert_file, e))
-    if now is None:
-        now = datetime.datetime.now()
-    return (cert.not_valid_after - now).days
-
-
 def set_crypto_backend(module):
     '''
     Sets which crypto backend to use (default: auto detection).
@@ -984,9 +542,7 @@ def set_crypto_backend(module):
     elif backend == 'openssl':
         HAS_CURRENT_CRYPTOGRAPHY = False
     elif backend == 'cryptography':
-        try:
-            cryptography.__version__
-        except Exception as dummy:
+        if not _ORIGINAL_HAS_CURRENT_CRYPTOGRAPHY:
             module.fail_json(msg=missing_required_lib('cryptography'))
         HAS_CURRENT_CRYPTOGRAPHY = True
     else:
@@ -1046,3 +602,22 @@ def handle_standard_module_arguments(module, needs_acme_v2=False):
     locale.setlocale(locale.LC_ALL, 'C')
 
     return backend
+
+
+def get_backend(module):
+    if HAS_CURRENT_CRYPTOGRAPHY:
+        return CryptographyBackend(module)
+    else:
+        return OpenSSLCLIBackend(module)
+
+
+def openssl_get_csr_identifiers(openssl_binary, module, csr_filename, csr_content=None):
+    return OpenSSLCLIBackend(module, openssl_binary=openssl_binary).get_csr_identifiers(csr_filename=csr_filename, csr_content=csr_content)
+
+
+def cryptography_get_csr_identifiers(module, csr_filename, csr_content=None):
+    return CryptographyBackend(module).get_csr_identifiers(csr_filename=csr_filename, csr_content=csr_content)
+
+
+def cryptography_get_cert_days(module, cert_file, now=None):
+    return CryptographyBackend(module).get_cert_days(cert_filename=cert_file, now=now)
